MASIVO2023020001-01-Creación-de-CRUD-criterio

+++ src/modules/kilometers/modules/criterios/criterios.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { CreateCriterioDto } from './dto/create-criterio.dto';
 import { UpdateCriterioDto } from './dto/update-criterio.dto';
 import { Criterio } from './entities/criterio.entity';
@@ -9,13 +9,9 @@ import { InjectRepository } from '@nestjs/typeorm';
 export class CriteriosService {
 
   constructor(
-    @InjectRepository(Criterio) private criterioRepository: Repository<Criterio>
+    @InjectRepository(Criterio) 
+    private criterioRepository: Repository<Criterio>
   ){}
-
-  create(createCriterioDto: CreateCriterioDto) {
-    const data = this.criterioRepository.create(createCriterioDto)
-    return this.criterioRepository.save(data)
-  }
 
   findAll() {
     return this.criterioRepository.find()
@@ -25,12 +21,47 @@ export class CriteriosService {
     return await this.criterioRepository.findBy({ id })
   }
 
-  update(id: number, updateCriterioDto: UpdateCriterioDto) {
-    const update = this.criterioRepository.create(updateCriterioDto)
-    return this.criterioRepository.update({ id }, update)
+  async findByName(nombre: string) {
+    const criterios = this.criterioRepository.find({ where: { nombre } })
+    return criterios
   }
 
-  remove(id: number) {
-    return this.criterioRepository.delete({ id })
+  async create(createCriterioDto: CreateCriterioDto) {
+    const create = this.criterioRepository.create(createCriterioDto)
+    const criteriosFound = await this.findByName(create.nombre)
+    if (criteriosFound.length > 0) {
+      throw new BadRequestException('Ya existe un criterio con ese nombre')
+    }
+    return this.criterioRepository.save(create)
+  }
+
+  async update(id: number, updateCriterioDto: UpdateCriterioDto) {
+    const criteriosFound = await this.findOne(id)
+    if (criteriosFound.length < 1) {
+      throw new BadRequestException('No se encontró un criterio con ese ID')
+    }
+    const update = this.criterioRepository.create(updateCriterioDto)
+    try {
+      this.criterioRepository.update({ id }, update)
+      return updateCriterioDto
+    } catch (err) {
+      throw new BadRequestException(err)
+    }
+  }
+
+  async remove(id: number) {
+    const criteriosFound = await this.findOne(id)
+    if (criteriosFound.length < 1) {
+      throw new BadRequestException('No se encontró un criterio con ese ID');
+    }
+    try {
+      this.criterioRepository.delete({id})
+      return {
+        status: 200,
+        message: `Se ha borrado el criterio con ID: ${id}`
+      }
+    } catch (err) {
+      throw new BadRequestException(err)
+    }
   }
 }
